# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portal Trigo",
    "description": "Backend do Portal Trigo.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "Fora da trava de troca: o front precisa ler ``mustChangePassword``.",
        "operationId": "me_api_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Trocar Senha",
        "description": "Única rota de escrita liberada durante a trava — e a saída dela.\n\nDevolve cookie novo: o ``token_version`` mudou e o antigo acabou de ser\nrevogado, então sem isto quem trocou a senha seria expulso da própria tela.",
        "operationId": "trocar_senha_api_auth_change_password_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "tags": [
          "produtos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_products_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "empori",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Empori"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Pagesize"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/sync": {
      "post": {
        "tags": [
          "produtos"
        ],
        "summary": "Sincronizar",
        "description": "Dispara a carga a partir do banco do Protheus.\n\n200 e não 202: a carga roda no próprio request e o relatório por empresa\nvolta na resposta. Quando o volume exigir execução em segundo plano, isto\nvira um job e a rota passa a devolver 202 com um id de acompanhamento.",
        "operationId": "sincronizar_api_products_sync_post",
        "parameters": [
          {
            "name": "empresa",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Empresa"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSyncResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{produto_id}": {
      "get": {
        "tags": [
          "produtos"
        ],
        "summary": "Obter",
        "description": "Rota com parâmetro vem DEPOIS de ``/sync``.\n\nRegistrada antes, ``sync`` seria lido como id de produto e a carga viraria\num 404 sem explicação.",
        "operationId": "obter_api_products__produto_id__get",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Produto Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProduct"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordInput": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1,
            "title": "Currentpassword"
          },
          "newPassword": {
            "type": "string",
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "title": "ChangePasswordInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginInput": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 1,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "Senha em texto, sobre TLS"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginInput",
        "description": "Credenciais do login.\n\nO e-mail NÃO passa por validação de formato aqui, e isso é deliberado. No\nlogin ele é apenas a chave de busca: recusar antes de consultar não protege\nnada — a senha ainda precisaria conferir — e o validador do Pydantic é mais\nrígido que o do backend anterior. Ele recusa domínio de uso interno\n(``.local``, por exemplo), que o ``zod.email()`` aceitava. Manter o rigor\naqui trancaria para fora qualquer conta em domínio interno já cadastrada.\n\nFormato de e-mail se valida no CADASTRO do usuário, onde recusar é barato e\na pessoa pode corrigir. Ver ``CriarUsuarioInput``."
      },
...
      "ProductListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PublicProduct"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "title": "ProductListResponse"
      },
      "ProductSyncCompanyResult": {
        "properties": {
          "empresa": {
            "type": "string",
            "title": "Empresa"
          },
          "filial": {
            "type": "string",
            "title": "Filial"
          },
          "empori": {
            "type": "string",
            "title": "Empori"
          },
          "sourceTable": {
            "type": "string",
            "title": "Sourcetable"
          },
          "paginas": {
            "type": "integer",
            "title": "Paginas"
          },
          "lidos": {
            "type": "integer",
            "title": "Lidos"
          },
          "gravados": {
            "type": "integer",
            "title": "Gravados"
          },
          "duplicados": {
            "type": "integer",
            "title": "Duplicados",
            "default": 0
          },
          "erro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Erro"
          }
        },
        "type": "object",
        "required": [
          "empresa",
          "filial",
          "empori",
          "sourceTable",
          "paginas",
          "lidos",
          "gravados"
        ],
        "title": "ProductSyncCompanyResult"
      },
      "ProductSyncResult": {
        "properties": {
          "empresas": {
            "items": {
              "$ref": "#/components/schemas/ProductSyncCompanyResult"
            },
            "type": "array",
            "title": "Empresas"
          },
          "lidos": {
            "type": "integer",
            "title": "Lidos"
          },
          "gravados": {
            "type": "integer",
            "title": "Gravados"
          },
          "duracaoMs": {
            "type": "integer",
            "title": "Duracaoms"
          }
        },
        "type": "object",
        "required": [
          "empresas",
          "lidos",
          "gravados",
          "duracaoMs"
        ],
        "title": "ProductSyncResult"
      },
...
      "PublicProduct": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "empori": {
            "type": "string",
            "title": "Empori"
          },
          "sourceTable": {
            "type": "string",
            "title": "Sourcetable"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "defaultWarehouse": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultwarehouse"
          },
          "ncm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncm"
          },
          "fiscalModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiscalmodel"
          },
          "isBlocked": {
            "type": "boolean",
            "title": "Isblocked"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "costCenter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Costcenter"
          },
          "expenseAccount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expenseaccount"
          },
          "assetAccount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetaccount"
          },
          "revenueAccount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenueaccount"
          },
          "salePrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saleprice"
          },
          "syncedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Syncedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "empori",
          "sourceTable",
          "code",
          "description",
          "type",
          "unit",
          "group",
          "defaultWarehouse",
          "ncm",
          "fiscalModel",
          "isBlocked",
          "isActive",
          "costCenter",
          "expenseAccount",
          "assetAccount",
          "revenueAccount",
          "salePrice",
          "syncedAt"
        ],
        "title": "PublicProduct"
      },
...
      "SessionResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/SessionUser"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "SessionResponse"
      },
      "SessionUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "mustChangePassword": {
            "type": "boolean",
            "title": "Mustchangepassword"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "mustChangePassword"
        ],
        "title": "SessionUser"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}